removed dialer pool
